refactor: Move inversePrecedence tests to separate test class
Extract lockInverseOrderWithLabel and lockInverseOrderMixedDifferentJobs
into LockStepInversePrecedenceTest to keep LockStepTest smaller and
avoid CI timeouts.
